table 의 pk 는 모두 id로 수정

+++ src/main/java/com/github/bestheroz/demo/api/admin/menu/AdminMenuController.java
@@ -48,13 +48,13 @@ public class AdminMenuController {
 
   @PutMapping(value = "{id}")
   public ResponseEntity<ApiResult> put(
-      @PathVariable(value = "id") final Integer id,
+      @PathVariable(value = "id") final Long id,
       @RequestBody final TableMenuEntity tableMenuEntity) {
     return Result.ok(this.adminMenuService.put(tableMenuEntity, id));
   }
 
   @DeleteMapping(value = "{id}")
-  public ResponseEntity<ApiResult> delete(@PathVariable(value = "id") final Integer id) {
+  public ResponseEntity<ApiResult> delete(@PathVariable(value = "id") final Long id) {
     return Result.ok(this.adminMenuService.delete(id));
   }
