+++ src/main/java/com/github/bestheroz/demo/api/entity/menu/TableMenuRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TableMenuRepository extends CrudRepository<TableMenuEntity, Integer> {
+public interface TableMenuRepository extends CrudRepository<TableMenuEntity, Long> {
   List<TableMenuEntity> findAll(Sort sort);
 }
